fix: give runtime CLI probes docker startup budget (#715)
* fix: give runtime CLI probes docker startup budget

Split the runtime-image CLI probe timeout from the fast provider/auth probe timeout so create-time Codex readiness does not fail on slow throwaway Docker startup while auth probes remain quick.

Validation: pytest tests/unit/service/test_provider_readiness_parts/test_provider_readiness_part_001.py tests/unit/service/test_provider_readiness_parts/test_provider_readiness_part_003.py tests/unit/service/test_provider_readiness_parts/test_provider_readiness_part_007.py -q; ruff check touched files; ruff format --check touched files; mypy touched provider readiness files; real selected-provider preflight returns ready with probe_detail=/usr/bin/codex.

* fix(ci): line-limit check — split provider readiness tests

---------
